completed MVP of social media application

+++ Threads-Clone/Frontend/src/hooks/useDeletePost.js
@@ -1,18 +1,24 @@
 import { useNavigate } from "react-router-dom";
-import { useRecoilValue } from "recoil";
+import { useRecoilState, useRecoilValue } from "recoil";
+import postsAtom from "../atoms/postsAtom";
 import userAtom from "../atoms/userAtom";
 import useShowToast from "./useShowToast";
 
-const useDeletePost = (post) => {
+const useDeletePost = () => {
   const showToast = useShowToast();
   const navigate = useNavigate();
   const currentUser = useRecoilValue(userAtom);
-  const handleDeletePost = async (e) => {
+  const [ posts, setPosts ] = useRecoilState(postsAtom);
+  const handleDeletePost = async (e, post) => {
     try {
+      if(!post){
+        return null;
+      }
       e.preventDefault();
       if (!window.confirm("Are you sure you want to delete this post?")) {
         return;
       }
+      console.log("useDeletePost post", post);
       const res = await fetch(`/api/posts/${post._id}`, {
         method: "DELETE",
       });
@@ -24,6 +30,11 @@ const useDeletePost = (post) => {
           showToast("Error", data.error, "error");
           return;
         }
+      }
+      if(Array.isArray(posts)){
+        setPosts(posts.filter((p) => p._id !== post._id));
+      } else {
+        setPosts(null);
       }
       showToast("Success", "Post Deleted", "success");
       navigate(`/${currentUser.username}`);
